매도 후 기존의 계좌 반영 누락을 고침 +진예슬 기능완료 merge

+++ src/main/java/com/project/nmt/service/OrderServiceImpl.java
@@ -2,7 +2,6 @@ package com.project.nmt.service;
 
 import java.time.LocalDate;
 
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -15,7 +14,6 @@ import com.project.nmt.repository.OrderLogRepository;
 import com.project.nmt.repository.OrderRepository;
 import com.project.nmt.repository.UserRepository;
 
-
 @Service
 public class OrderServiceImpl implements OrderService {
 	@Autowired
@@ -24,26 +22,29 @@ public class OrderServiceImpl implements OrderService {
 	UserRepository userRepository;
 	@Autowired
 	OrderLogRepository orderLogRepository;
-	
+
 	@Override
 	public void buy(User user, StockInfo stockInfo, Order order, Long totalPrice, int cnt) {
-		//매수한 금액만큼 계좌에서 차감
-		int updateBudget=(int) (user.getBudget()-totalPrice);
+		// 매수한 금액만큼 계좌에서 차감
+		int updateBudget = (int) (user.getBudget() - totalPrice);
 		userRepository.updateBudget(updateBudget, user.getId());
-		
-		int totalCnt=order.getQuantity();
-		int orderTotalPrice=order.getPrice()*totalCnt;
-		//매수후 평가금액
-		int afterAvgPrice=(int) ((orderTotalPrice+totalPrice)/(totalCnt+cnt));
-		
-		//기존의 order 삭제
-		orderRepository.deleteById(order.getId());
-		
-		//이전 주문내역과 현재 매수로 갱신한 새로운 order삽입
+		int afterAvgPrice=(int) (totalPrice/cnt);
+		int totalCnt=cnt;
+		if (order != null) {
+			totalCnt = order.getQuantity();
+			int orderTotalPrice = order.getPrice() * totalCnt;
+			// 매수후 평가금액
+			afterAvgPrice = (int) ((orderTotalPrice + totalPrice) / (totalCnt + cnt));
+
+			// 기존의 order 삭제
+			orderRepository.deleteById(order.getId());
+		}
+
+		// 이전 주문내역과 현재 매수로 갱신한 새로운 order삽입
 		Order newOrder = new Order();
 		newOrder.setBoughtDate(LocalDate.now());
 		newOrder.setPrice(afterAvgPrice);
-		newOrder.setQuantity(cnt+totalCnt);
+		newOrder.setQuantity(cnt + totalCnt);
 		newOrder.setStock(stockInfo.getStock());
 		newOrder.setUser(user);
 		orderRepository.save(newOrder);
@@ -51,18 +52,17 @@ public class OrderServiceImpl implements OrderService {
 
 	@Override
 	public void sell(User user, StockInfo stockInfo, Stock nowStock, Order order, Long totalPrice, int cnt) {
-		if(order.getQuantity()==cnt) {//판매 개수와 보유 개수가 같으면 삭제
+		if (order.getQuantity() == cnt) {// 판매 개수와 보유 개수가 같으면 삭제
 			orderRepository.deleteOrderById(order.getId());
-		}
-		else {//보유개수가 더 많으면 차감
-			orderRepository.updateQuantityById(order.getId(), order.getQuantity()-cnt);
+		} else {// 보유개수가 더 많으면 차감
+			orderRepository.updateQuantityById(order.getId(), order.getQuantity() - cnt);
 		}
 
-       //현재 가격 기준 판매한 금액만큼 계좌에 합산
-		userRepository.updateBudget(stockInfo.getPrice()*cnt, user.getId());
-		
-		//판것에 대해 log
-		OrderLog orderLog=new OrderLog();
+		// 현재 가격 기준 판매한 금액만큼 계좌에 합산
+		userRepository.updateBudget(stockInfo.getPrice() * cnt+user.getBudget(), user.getId());
+
+		// 판것에 대해 log
+		OrderLog orderLog = new OrderLog();
 		orderLog.setSoldDate(LocalDate.now());
 		orderLog.setSoldPrice(stockInfo.getPrice());
 		orderLog.setBoughtPrice(order.getPrice());
